Revert "CharacterCollisionHandler"
