--- Assets/Scripts/World/LettersToReid/CharacterCollisonHandler.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 303632d2a7a294c98b6c1b71e502f7c6
-timeCreated: 1487443689
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
