+++ Assets/Scripts/World/LettersToReid/MyCharacterController.cs
@@ -10,9 +10,12 @@ public class MyCharacterController : MonoBehaviour {
     public float climbSpeed = .1f;
     public bool isGrounded;
     public bool canClimb = false;
-    
+    public int  enemyHits = 0;
 
-    
+    Vector3[] memoryLocations = { new Vector3(5.76f, 0.26f, 0), new Vector3(16.21f, 2.37f, 0),
+    new Vector3(7.1f, 2.43f, 0), new Vector3(-6.06f, 2.43f, 0), new Vector3(-10.66f, 4.55f, 0),
+    new Vector3(-16.82f, 6.74f, 0), new Vector3(3.77f, 6.69f, 0), new Vector3(11.13f, 4.56f, 0),
+    new Vector3(13.46f, 8.91f, 0), new Vector3(15.87f, 11.08f, 0)};
 
     // Use this for initialization
     void Start () {
@@ -67,5 +70,38 @@ public class MyCharacterController : MonoBehaviour {
         }
     }
 
-   
+    void OnCollisionEnter2D(Collision2D col)
+    {
+        if(col.collider.tag == "ground")
+        {
+            isGrounded = true;
+        }
+
+        else if (col.gameObject.tag == "ladder")
+        {
+            canClimb = true;
+            Debug.Log("Enter");
+        }
+        else if (col.gameObject.tag == "enemy")
+        {
+            transform.position = memoryLocations[enemyHits];
+            if(enemyHits < 9 )
+            {
+                enemyHits++;
+            }
+            else if(enemyHits >= 9)
+            {
+                enemyHits = 0;
+            }
+        }
+    }
+
+    void OnCollisionExit2D(Collision2D col2)
+    {
+        if (col2.gameObject.tag == "ladder")
+        {
+                canClimb = false;
+                Debug.Log("Exit");
+        }
+    }
 }
